Migrate resource to the new app
