+++ common/management/commands/data_creation/create_reservation_units.py
@@ -28,10 +28,9 @@ from reservation_units.models import (
     TaxPercentage,
 )
 from reservations.models import ReservationMetadataSet
-from resources.models import Resource
 from spaces.models import Unit
 from tilavarauspalvelu.enums import TermsOfUseTypeChoices
-from tilavarauspalvelu.models import Service, TermsOfUse
+from tilavarauspalvelu.models import Resource, Service, TermsOfUse
 
 from .create_seasonal_booking import _create_application_round_time_slots
 from .utils import (
