--- resources/translation.py
@@ -1,8 +1,0 @@
-from modeltranslation.translator import TranslationOptions, register
-
-from resources.models import Resource
-
-
-@register(Resource)
-class ResourceTranslationOptions(TranslationOptions):
-    fields = ["name"]
