+++ tests/test_graphql_api/test_resource/test_create_permissions.py
@@ -1,8 +1,8 @@
 import pytest
 
-from resources.enums import ResourceLocationType
-from resources.models import Resource
 from tests.factories import SpaceFactory, UnitGroupFactory, UserFactory
+from tilavarauspalvelu.enums import ResourceLocationType
+from tilavarauspalvelu.models import Resource
 
 from .helpers import CREATE_MUTATION
