+++ tilavarauspalvelu/admin/__init__.py
@@ -1,6 +1,7 @@
 from .payment_accounting.admin import PaymentAccountingAdmin
 from .payment_merchant.admin import PaymentMerchantAdmin
 from .payment_order.admin import PaymentOrderAdmin
+from .resource.admin import ResourceAdmin
 from .service.admin import ServiceAdmin
 from .terms_of_use.admin import TermsOfUseAdmin
 from .user.admin import UserAdmin
@@ -9,6 +10,7 @@ __all__ = [
     "PaymentAccountingAdmin",
     "PaymentMerchantAdmin",
     "PaymentOrderAdmin",
+    "ResourceAdmin",
     "ServiceAdmin",
     "TermsOfUseAdmin",
     "UserAdmin",
