+++ tilavarauspalvelu/api/graphql/types/resource/filtersets.py
@@ -6,8 +6,8 @@ from graphene_django_extensions import ModelFilterSet
 from graphene_django_extensions.filters import IntMultipleChoiceFilter
 
 from permissions.enums import UserRoleChoice
-from resources.models import Resource
 from spaces.models import Space
+from tilavarauspalvelu.models import Resource
 
 if TYPE_CHECKING:
     from common.typing import AnyUser
