+++ tilavarauspalvelu/models/__init__.py
@@ -3,6 +3,7 @@ from .payment_merchant.model import PaymentMerchant
 from .payment_order.model import PaymentOrder
 from .payment_product.model import PaymentProduct
 from .personal_info_view_log.model import PersonalInfoViewLog
+from .resource.model import Resource
 from .service.model import Service
 from .terms_of_use.model import TermsOfUse
 from .user.model import ProfileUser, User
@@ -14,6 +15,7 @@ __all__ = [
     "PaymentProduct",
     "PersonalInfoViewLog",
     "ProfileUser",
+    "Resource",
     "Service",
     "TermsOfUse",
     "User",
