+++ tilavarauspalvelu/models/resource/actions.py
@@ -1,0 +1,9 @@
+from typing import TYPE_CHECKING
+
+if TYPE_CHECKING:
+    from .model import Resource
+
+
+class ResourceActions:
+    def __init__(self, resource: "Resource") -> None:
+        self.resource = resource
