+++ tilavarauspalvelu/translation.py
@@ -2,6 +2,7 @@ from modeltranslation.decorators import register
 from modeltranslation.translator import TranslationOptions
 
 from .models import Service, TermsOfUse
+from .models.resource.model import Resource
 
 
 @register(Service)
@@ -12,3 +13,8 @@ class ServiceTranslationOptions(TranslationOptions):
 @register(TermsOfUse)
 class TermsOfUseTranslationOptions(TranslationOptions):
     fields = ["name", "text"]
+
+
+@register(Resource)
+class ResourceTranslationOptions(TranslationOptions):
+    fields = ["name"]
